Added list of feature functions

+++ src/ca/ece/ubc/cpen221/mp5/statlearning/MP5Function.java
@@ -11,6 +11,6 @@ public interface MP5Function {
 	 * @param yelpRestaurant
 	 * @return the value of the feature function
 	 */
-	public int f(Restaurant yelpRestaurant, RestaurantDB db);
+	public double f(Restaurant yelpRestaurant, RestaurantDB db);
 	
 }
